feat(i18n): translate application to English and improve ingredient form
- feat: translate login component to English
- feat: add MeasurementUnit enum for standardized units
- refactor: update ingredient forms to use MeasurementUnit enum
- chore: update environment configuration

+++ RecipeMaster-API/RecipeMaster.Application/DTOs/UpdateIngredientDTO.cs
@@ -6,8 +6,8 @@ public class UpdateIngredientDTO
     public string Description { get; set; }
     public string Unit { get; set; }
     public decimal Cost { get; set; }
-    public decimal StockQuantity { get; set; }
-    public decimal MinimumStockLevel { get; set; }
+    public int StockQuantity { get; set; }
+    public int MinimumStockLevel { get; set; }
     public string SupplierName { get; set; }
     public bool IsPerishable { get; set; }
     public string OriginCountry { get; set; }
